abstracted ta catalog management an downloading into custom hook
